activate run button for active (but not necessarily selected), executable files
The run button will become active if an executable file is either selected, or active
(i.e. it is opened in the currently selected tab). The selected file takes precedence
over the currently active file: when clicked, the selected file will be executed, and
only if no file is selected, the active file will be executed, instead.